move notifier class to providers
